сделал хорошую(по-моему мнению) архитектуру кода

+++ pkg/post/post.go
@@ -7,28 +7,20 @@ import (
 	"time"
 )
 
-func Page() *models.IndexResp {
-	Resp := new(models.IndexResp)
-	Resp.Index = "Add Post"
-	Resp.Title = "Add"
-	return Resp
-}
 
-func Add(db *sql.DB, text string) error {
+func Add(db *sql.DB, userID int, text string) error {
 	year, month, day := time.Now().Date()
 	hour, min, _ := time.Now().Clock()
 	date:=fmt.Sprintf("%d:%d %d %s %d", hour, min, day, month, year)
-	fmt.Println(date)
-	fmt.Println(text)
-	_, err := db.Exec("INSERT INTO posts (body, time_date) VALUES ($1, $2)", text, date)
+	_, err := db.Exec("INSERT INTO posts (user_id,body, time_date) VALUES ($1, $2,$3)", userID,text, date)
 	if err != nil {
 		return err
 	}
 	return nil
 }
 
-func Posts(db *sql.DB) ([]*models.Post,error) {
-	rows, err :=db.Query("SELECT * FROM posts")
+func Posts(db *sql.DB, userID int) ([]*models.Post,error) {
+	rows, err :=db.Query("SELECT * FROM posts WHERE user_id=$1", userID)
 	if err != nil {
 		return nil, err
 	}
@@ -36,7 +28,7 @@ func Posts(db *sql.DB) ([]*models.Post,error) {
 	defer rows.Close()
 	for rows.Next(){
 		post := new(models.Post)
-		err := rows.Scan(&post.ID, &post.Body, &post.Date)
+		err := rows.Scan(&post.ID, &post.UserID,&post.Body, &post.Date)
 		if err != nil {
 			return nil, err
 		}
@@ -48,8 +40,8 @@ func Posts(db *sql.DB) ([]*models.Post,error) {
 	return posts, nil
 }
 
-func DeletePost(db *sql.DB, ID string) error {
-	_, err := db.Exec("DELETE FROM  posts WHERE id=$1", ID)
+func DeletePost(db *sql.DB, ID string, userID int) error {
+	_, err := db.Exec("DELETE FROM  posts WHERE id=$1 AND user_id=$2", ID, userID)
 	if err != nil {
 		return err
 	}
